optimisation/requirements.txt
scikit-learn
requests
flask
flask-cors
azure-storage-blob
gunicorn
